In the noon.. Only minor bugs
